services/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.128.7,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "neo4j (>=6.1.0,<7.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "app (>=0.0.1,<0.0.2)",
    "anthropic (>=0.84.0,<0.85.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "faker (>=40.5.1,<41.0.0)",
    "xgboost (>=3.2.0,<4.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "torch (>=2.10.0,<3.0.0) ; python_version >= \"3.12\" and python_version < \"3.15\"",
    "torch-geometric (>=2.7.0,<3.0.0) ; python_version >= \"3.12\" and python_version < \"3.15\"",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "langgraph (>=1.0.10,<2.0.0) ; python_version >= \"3.12\" and python_full_version < \"4.0.0\"",
    "langchain (>=1.2.10,<2.0.0) ; python_version >= \"3.12\" and python_full_version < \"4.0.0\"",
    "langchain-anthropic (>=1.3.4,<2.0.0) ; python_version >= \"3.12\" and python_full_version < \"4.0.0\"",
    "langchain-openai (>=1.1.10) ; python_version >= \"3.12\" and python_full_version < \"4.0.0\"",
    "langsmith (>=0.7.12,<0.8.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "pypdf (>=6.7.5,<7.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
]

[tool.poetry]
package-mode = false
packages = [
    { include = "data_gen", from = "src" },
    { include = "graph", from = "src" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "uvicorn[standard] (>=0.40.0,<0.41.0)"
]
